refactor: move updateSectionWarnings function to `updateWarnings.ts`
